+++ webauthin/views.py
@@ -10,9 +10,11 @@ from django.contrib.auth import login
 from django.contrib.auth.decorators import login_required
 from django.contrib.sites.shortcuts import get_current_site
 from django.http import JsonResponse
+from django.shortcuts import redirect
 from django.shortcuts import render
 from django.views.decorators.csrf import csrf_exempt
 
+from . import settings as wa_settings
 from .models import AuthData
 
 ICON = "https://example.com/"
@@ -78,7 +80,7 @@ def register_verify(request):
         webauthn_credential = webauthn_registration_response.verify()
     except Exception as e:
         messages.error(request, "Registration failed. Error: {}".format(e))
-        return JsonResponse({"fail": "Registration failed. Error: {}".format(e)})
+        return redirect(wa_settings.REGISTRATION_ERROR_URL)
 
     auth_data = AuthData.objects.filter(credential_id=webauthn_credential.credential_id)
     if auth_data.exists():
@@ -86,7 +88,7 @@ def register_verify(request):
             request,
             "This key is already registered to an account. Try logging in with it.",
         )
-        return JsonResponse({"fail": "Credential ID already exists."}, status=422)
+        return redirect(wa_settings.REGISTRATION_ERROR_URL)
 
     webauthn_credential.credential_id = str(webauthn_credential.credential_id, "utf-8")
     webauthn_credential.public_key = str(webauthn_credential.public_key, "utf-8")
@@ -96,7 +98,7 @@ def register_verify(request):
         public_key=webauthn_credential.public_key,
     )
     messages.success(request, "Your key has been successfully registered.")
-    return JsonResponse({"success": "User successfully registered."})
+    return redirect(wa_settings.REGISTRATION_REDIRECT_URL)
 
 
 @csrf_exempt
@@ -127,13 +129,14 @@ def login_verify(request):
     challenge = request.session.get("challenge")
     if not challenge:
         messages.error(request, "No challenge exists for your session.")
-        return JsonResponse("No challenge exists for your session.", status=422)
+        return redirect(wa_settings.LOGIN_ERROR_URL)
 
     user = authenticate(request, credential_id=request.POST["id"], data=request.POST)
     if user is None:
-        messages.error(request, "There was an error while validating your credentials.")
-        return JsonResponse({"fail": "Assertion failed."}, status=422)
+        messages.error(request, "Your credentials could not be validated.")
+        return redirect(wa_settings.LOGIN_ERROR_URL)
 
     login(request, user)
 
-    return JsonResponse({"success": True})
+    messages.success(request, "You have been successfully logged in.")
+    return redirect(wa_settings.LOGIN_REDIRECT_URL)