+++ CHANGELOG.md
@@ -9,4 +9,8 @@ Features
 ~~~~~~~~
 - Add alerts on registration errors. [Stavros Korokithakis]
 
+Fixes
+~~~~~
+- Improve error messages. [Stavros Korokithakis]
+